+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.8.10 Live (January 30, 2025)
 
+* Added a parallaxing reflection below the hero image (Demo).
 * Fixes to some of the HTML errors that the validator picked up. WordPress itself seems to generate many of them.
 * More adjustments and bug fixes.